+++ SavingsManager.csproj
@@ -226,6 +226,7 @@
     <Compile Include="App_Start\RouteConfig.cs" />
     <Compile Include="App_Start\Startup.Auth.cs" />
     <Compile Include="Controllers\GroupController.cs" />
+    <Compile Include="Controllers\SocioController.cs" />
     <Compile Include="Factory\ISavingsProvider.cs" />
     <Compile Include="Factory\SavingsProviderFactory.cs" />
     <Compile Include="Controllers\AccountController.cs" />
@@ -372,9 +373,9 @@
     <Content Include="Views\Group\NuevoGrupo.cshtml" />
     <Content Include="Views\Group\VerGrupos.cshtml" />
     <Content Include="Views\Group\EditarGrupo.cshtml" />
-    <Content Include="Views\Home\NuevoSocio.cshtml" />
-    <Content Include="Views\Home\VerSocios.cshtml" />
-    <Content Include="Views\Home\EditarSocio.cshtml" />
+    <Content Include="Views\Socio\NuevoSocio.cshtml" />
+    <Content Include="Views\Socio\VerSocios.cshtml" />
+    <Content Include="Views\Socio\EditarSocio.cshtml" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="App_Data\" />
