+++ Views/Shared/_Layout.cshtml
@@ -40,8 +40,8 @@
                     <li>
                         Asociados
                         <ul>
-                            <li>@Html.ActionLink("Registrar Nuevo Asociado", "NuevoSocio", "Home")</li>
-                            <li>@Html.ActionLink("Ver Asociados", "VerSocios", "Home")</li> 
+                            <li>@Html.ActionLink("Registrar Nuevo Asociado", "NuevoSocio", "Socio")</li>
+                            <li>@Html.ActionLink("Ver Asociados", "VerSocios", "Socio")</li> 
                         </ul>
                     </li>
                     <li>
